requirements.txt
streamlit
docx-revisions
python-docx
lxml
